+++ test/large_tests/verified_integer.cpp
@@ -149,7 +149,7 @@ void VerifiedInteger::FindAll(Array &c, int64_t value, size_t start, size_t end)
 	if (c.Size() != result.size())
 		assert(false);
 	for(size_t t = 0; t < result.size(); ++t) {
-		if (result[t] != c.Get(t)) 
+		if (result[t] != (size_t)c.Get(t)) 
 			assert(false);
 	}
